refactor: deepen architecture — 4 structural improvements
01: Split models.ts (445 lines) into catalog + discovery + thinking
    - src/thinking.ts: reasoning-effort maps + translation interface
    - src/catalog.ts: static model data + compat + filtering
    - src/discovery.ts: fetch / parse / resolveModels
    - src/models.ts: barrel re-export (no consumer changes)

02: Deep thinking module — reasoningEffortFor() owns the translation
    interface; both catalog and discovery use the same maps.

03: Collapse walkAuthPaths into resolveApiKey
    - Removed generic <T>/extract seam with one caller
    - All walk logic inlined; injectable I/O handles kept

04: Clean env.ts — moved sanitizeApiKey to oauth.ts (only caller),
    deleted buildEndpointUrl + DEFAULT_ENDPOINT (dead code)

Win: model data edits now touch only catalog.ts (138 lines),
not the 445-line hot spot. Compat-merge and map-selection rules
stated once each. 97 tests pass across 9 files.
