+++ src/auth.ts
@@ -6,8 +6,6 @@
  * 2. QWENCLOUD_API_KEY environment variable
  * 3. pi auth config files (~/.pi/agent/auth.json)
  *
- * Unlike ClinePass, QwenCloud has no OAuth/WorkOS flow — only static API keys.
- *
  * @module qwencloud-auth
  */
 
@@ -17,9 +15,7 @@ import { join } from "node:path";
 import { isRecord, stringValue } from "./utils.js";
 import { ENV_API_KEY, PROVIDER_NAME } from "./env.js";
 
-/**
- * Default auth file paths checked in order.
- */
+/** Default auth file paths checked in order. */
 export function defaultAuthPaths(home: string): string[] {
   return [join(home, ".pi", "agent", "auth.json")];
 }
@@ -33,13 +29,24 @@ export interface AuthKeyOptions {
 }
 
 /**
- * Iterate auth file paths in order, parsing JSON from each and extracting
- * a value. Suppresses ENOENT, warns on corrupt files.
+ * Resolve the QwenCloud API key.
+ * Priority: provided key → QWENCLOUD_API_KEY env var → auth files
+ *
+ * Auth files checked:
+ * - ~/.pi/agent/auth.json:
+ *   { "apiKey": "<key>" }
+ *   { "qw": "<key>" } or { "qw": { "access": "<key>" } }
  */
-export function walkAuthPaths<T>(
-  options: AuthKeyOptions,
-  extract: (parsed: Record<string, unknown>) => T | undefined,
-): T | undefined {
+export function resolveApiKey(
+  providedKey?: string,
+  options: AuthKeyOptions = {},
+): string | undefined {
+  if (providedKey) return providedKey;
+
+  const env = options.env ?? process.env;
+  if (env[ENV_API_KEY]) return env[ENV_API_KEY];
+
+  // Walk auth file paths — formerly the generic walkAuthPaths<T> seam.
   const home = options.homeDir?.() ?? homedir();
   const authPaths = options.authPaths ?? defaultAuthPaths(home);
   const readFile = options.readFile ?? ((p: string) => readFileSync(p, "utf-8"));
@@ -51,8 +58,17 @@ export function walkAuthPaths<T>(
       const parsed: unknown = JSON.parse(readFile(authPath));
       if (!isRecord(parsed)) continue;
 
-      const result = extract(parsed);
-      if (result !== undefined) return result;
+      // Direct apiKey field
+      const apiKey = stringValue(parsed.apiKey);
+      if (apiKey) return apiKey;
+
+      // qw field (string or OAuth-style object)
+      const qcField = parsed[PROVIDER_NAME];
+      if (typeof qcField === "string") return qcField;
+      if (isRecord(qcField)) {
+        const access = stringValue(qcField.access);
+        if (access) return access;
+      }
     } catch (e) {
       const msg = e instanceof Error ? e.message : String(e);
       if (!msg.includes("ENOENT") && !msg.includes("not found")) {
@@ -61,38 +77,4 @@ export function walkAuthPaths<T>(
     }
   }
   return undefined;
-}
-
-/**
- * Resolve the QwenCloud API key.
- * Priority: provided key → QWENCLOUD_API_KEY env var → auth files
- *
- * Auth files checked:
- * - ~/.pi/agent/auth.json:
- *   { "qwencloud": "<key>" } or { "qwencloud": { "access": "<key>" } }
- *   or { "apiKey": "<key>" }
- */
-export function resolveApiKey(
-  providedKey?: string,
-  options: AuthKeyOptions = {},
-): string | undefined {
-  if (providedKey) return providedKey;
-
-  const env = options.env ?? process.env;
-  if (env[ENV_API_KEY]) return env[ENV_API_KEY];
-
-  return walkAuthPaths(options, (parsed) => {
-    // Direct apiKey field
-    const apiKey = stringValue(parsed.apiKey);
-    if (apiKey) return apiKey;
-
-    // qwencloud field (string or OAuth-style object)
-    const qcField = parsed[PROVIDER_NAME];
-    if (typeof qcField === "string") return qcField;
-    if (isRecord(qcField)) {
-      const access = stringValue(qcField.access);
-      if (access) return access;
-    }
-    return undefined;
-  });
 }
